// manager_frontend/src/client/models/DashboardPeriod.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DashboardPeriodRange } from './DashboardPeriodRange';
export type DashboardPeriod = {
    kind?: string;
    timezone?: string;
    current: DashboardPeriodRange;
    previous: DashboardPeriodRange;
};